refactor: 댓글 이동 커스텀 훅 생성
